fix: Resolve remaining TypeScript type errors

+++ packages/core/src/logic/smart-extraction-with-circuit-breaker.ts
@@ -64,7 +64,7 @@ export class ResilientExtractionOrchestrator {
 
         return {
           success: true,
-          amount: result,
+          amount: result as bigint,
           methodUsed: methodName,
         }
       } catch (error) {
